backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
transformers==4.40.0
torch==2.2.1
pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
huggingface-hub==0.21.4
timm==0.9.16